+++ tests/tools/pylint/pylint_configs/multiple/testcode.py
@@ -1,0 +1,3 @@
+def name_longer_than_40_chars_but_max_is_40_so_should_error():
+    this_is_a_bad_name = 5
+    print(this_is_a_bad_name)