+++ tests/tools/pylint/test_pylint_tool.py
@@ -7,6 +7,8 @@ from prospector.config import ProspectorConfig
 from prospector.finder import FileFinder
 from prospector.tools.pylint import PylintTool
 
+THIS_DIR = Path(__file__).parent
+
 
 def _get_pylint_tool_and_prospector_config(argv_patch=None):
     if argv_patch is None:
@@ -18,11 +20,40 @@ def _get_pylint_tool_and_prospector_config(argv_patch=None):
 
 
 def _get_test_files(*names: str, workdir: Path = None):
-    paths = [Path(__file__).parent / name for name in names]
+    paths = [THIS_DIR / name for name in names]
     return FileFinder(*paths, workdir=workdir)
 
 
 class TestPylintTool(TestCase):
+    def test_pylint_config(self):
+        """Verifies that prospector will configure pylint with any pylint-specific configuration if found"""
+
+        def _has_message(msg_list, code):
+            return any([message.code == code and message.source == "pylint" for message in msg_list])
+
+        for config_type in ("pylintrc", "pylintrc2", "pyproject", "setup.cfg"):
+            root = THIS_DIR / "pylint_configs" / config_type
+
+            with patch("os.getcwd", return_value=root.absolute()):
+                pylint_tool, config = _get_pylint_tool_and_prospector_config()
+            self.assertEqual(Path(config.workdir).absolute(), root.absolute())
+
+            found_files = _get_test_files(root)
+            pylint_tool.configure(config, found_files)
+
+            messages = pylint_tool.run(found_files)
+            self.assertTrue(_has_message(messages, "line-too-long"), msg=config_type)
+
+    def test_absolute_path_is_computed_correctly(self):
+        pylint_tool, config = _get_pylint_tool_and_prospector_config()
+        root = os.path.join(os.path.dirname(__file__), "testpath", "test.py")
+        root_sep_split = root.split(os.path.sep)
+        root_os_split = os.path.split(root)
+        found_files = _get_test_files("testpath/test.py")
+        pylint_tool.configure(config, found_files)
+        self.assertNotEqual(pylint_tool._args, [os.path.join(*root_sep_split)])
+        self.assertEqual(pylint_tool._args, [os.path.join(*root_os_split)])
+
     def test_wont_throw_false_positive_relative_beyond_top_level(self):
         with patch("os.getcwd", return_value=os.path.realpath("tests/tools/pylint/testpath/")):
             pylint_tool, config = _get_pylint_tool_and_prospector_config()
@@ -49,4 +80,16 @@ class TestPylintTool(TestCase):
         )
         found_files = _get_test_files("testpath/absolute-import/pkg", workdir=Path(workdir))
         pylint_tool.configure(config, found_files)
-        pylint_tool.run(found_files)
+        messages = pylint_tool.run(found_files)
+        self.assertEqual(messages[0].code, "no-name-in-module")
+
+    def test_use_prospector_default_path_finder(self):
+        workdir = "tests/tools/pylint/testpath/absolute-import/"
+        with patch("os.getcwd", return_value=os.path.realpath(workdir)):
+            pylint_tool, config = _get_pylint_tool_and_prospector_config(
+                argv_patch=["", "-P", "prospector-default-finder"]
+            )
+        found_files = _get_test_files("testpath/absolute-import/pkg", workdir=Path(workdir))
+        pylint_tool.configure(config, found_files)
+        messages = pylint_tool.run(found_files)
+        self.assertEqual(messages[0].code, "no-name-in-module")